# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Svarog - Solana Token Launcher Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard overview stats
      parameters:
        - name: walletAddress
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/activity:
    get:
      operationId: getDashboardActivity
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: walletAddress
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"

  # Wallets
  /wallets:
    get:
      operationId: listWallets
      tags: [wallets]
      summary: List all stored wallets
      parameters:
        - name: ownerAddress
          in: query
          required: false
          schema:
            type: string
        - name: group
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wallet"
    post:
      operationId: createWallet
      tags: [wallets]
      summary: Generate a new wallet keypair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletInput"
      responses:
        "201":
          description: Created wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallets/import:
    post:
      operationId: importWallet
      tags: [wallets]
      summary: Import wallet by private key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletImport"
      responses:
        "201":
          description: Imported wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "400":
          description: Invalid private key
  /wallets/generate-bulk:
    post:
      operationId: generateBulkWallets
      tags: [wallets]
      summary: Generate multiple wallets at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkWalletInput"
      responses:
        "201":
          description: Generated wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Wallet"
  /wallets/{id}:
    get:
      operationId: getWallet
      tags: [wallets]
      summary: Get a specific wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "404":
          description: Not found
    patch:
      operationId: updateWallet
      tags: [wallets]
      summary: Update wallet metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletUpdate"
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
    delete:
      operationId: deleteWallet
      tags: [wallets]
      summary: Delete a wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /wallets/{id}/balance:
    get:
      operationId: getWalletBalance
      tags: [wallets]
      summary: Get wallet SOL balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
  /wallets/groups:
    get:
      operationId: listWalletGroups
      tags: [wallets]
      summary: List all wallet groups
      parameters:
        - name: ownerAddress
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WalletGroup"

  # Bundles
...
  /tokens/metadata:
    get:
      operationId: getTokenMetadata
      tags: [tokens]
      summary: Fetch token metadata from Pump.Fun by contract address
      parameters:
        - name: ca
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenMetadata"
        "404":
          description: Token not found

  # Settings
  /settings:
    get:
      operationId: getSettings
      tags: [dashboard]
      summary: Get user settings
      parameters:
        - name: walletAddress
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
    patch:
      operationId: updateSettings
      tags: [dashboard]
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
components:
  schemas:
...
    DashboardStats:
      type: object
      required: [earningsToday, earningsLast30Days, bundlesLaunched, totalBalanceSol, totalBalanceUsd, pnl]
      properties:
        earningsToday:
          type: number
        earningsLast30Days:
          type: number
        bundlesLaunched:
          type: integer
        totalBalanceSol:
          type: number
        totalBalanceUsd:
          type: number
        pnl:
          type: number
        activeBots:
          type: integer
        totalWallets:
          type: integer
    ActivityEvent:
      type: object
      required: [id, type, description, timestamp]
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        timestamp:
          type: string
        tokenName:
          type: ["string", "null"]
        tokenSymbol:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        txHash:
          type: ["string", "null"]
        walletAddress:
          type: ["string", "null"]

    # Wallets
    Wallet:
      type: object
      required: [id, publicKey, ownerAddress, label, group, createdAt]
      properties:
        id:
          type: integer
        publicKey:
          type: string
        ownerAddress:
          type: string
        label:
          type: ["string", "null"]
        group:
          type: ["string", "null"]
        balanceSol:
          type: ["number", "null"]
        balanceUsd:
          type: ["number", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    WalletInput:
      type: object
      required: [ownerAddress]
      properties:
        ownerAddress:
          type: string
        label:
          type: string
        group:
          type: string
    WalletImport:
      type: object
      required: [ownerAddress, privateKey]
      properties:
        ownerAddress:
          type: string
        privateKey:
          type: string
        label:
          type: string
        group:
          type: string
    BulkWalletInput:
      type: object
      required: [ownerAddress, count]
      properties:
        ownerAddress:
          type: string
        count:
          type: integer
        group:
          type: string
    WalletUpdate:
      type: object
      properties:
        label:
          type: string
        group:
          type: string
        isActive:
          type: boolean
    WalletBalance:
      type: object
      required: [publicKey, balanceSol]
      properties:
        publicKey:
          type: string
        balanceSol:
          type: number
        balanceUsd:
          type: ["number", "null"]
    WalletGroup:
      type: object
      required: [name, count]
      properties:
        name:
          type: string
        count:
          type: integer
        totalBalanceSol:
          type: ["number", "null"]

    # Bundles
...
    TokenMetadata:
      type: object
      required: [address, name, symbol]
      properties:
        address:
          type: string
        name:
          type: string
        symbol:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        decimals:
          type: ["integer", "null"]
        supply:
          type: ["number", "null"]

    # Settings
    UserSettings:
      type: object
      required: [walletAddress]
      properties:
        walletAddress:
          type: string
        defaultWalletCount:
          type: integer
        defaultSolPerWallet:
          type: number
        autoApprove:
          type: boolean
        darkMode:
          type: boolean
        notificationsEnabled:
          type: boolean
        rpcEndpoint:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    SettingsUpdate:
      type: object
      properties:
        walletAddress:
          type: string
        defaultWalletCount:
          type: integer
        defaultSolPerWallet:
          type: number
        autoApprove:
          type: boolean
        darkMode:
          type: boolean
        notificationsEnabled:
          type: boolean
        rpcEndpoint:
          type: string