# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Svarog - Solana Token Launcher Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityEvent {
  id: number;
  type: string;
  description: string;
  timestamp: string;
  /** @nullable */
  tokenName?: string | null;
  /** @nullable */
  tokenSymbol?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  walletAddress?: string | null;
}