# lib/api-zod/src/generated/types/bulkWalletInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Svarog - Solana Token Launcher Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BulkWalletInput {
  ownerAddress: string;
  count: number;
  group?: string;
}