# lib/api-zod/src/generated/types/settingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Svarog - Solana Token Launcher Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SettingsUpdate {
  walletAddress?: string;
  defaultWalletCount?: number;
  defaultSolPerWallet?: number;
  autoApprove?: boolean;
  darkMode?: boolean;
  notificationsEnabled?: boolean;
  rpcEndpoint?: string;
}