# lib/api-zod/src/generated/types/userSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Svarog - Solana Token Launcher Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserSettings {
  walletAddress: string;
  defaultWalletCount?: number;
  defaultSolPerWallet?: number;
  autoApprove?: boolean;
  darkMode?: boolean;
  notificationsEnabled?: boolean;
  /** @nullable */
  rpcEndpoint?: string | null;
  createdAt?: string;
  updatedAt?: string;
}