# lib/api-zod/src/generated/types/walletGroup.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Svarog - Solana Token Launcher Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface WalletGroup {
  name: string;
  count: number;
  /** @nullable */
  totalBalanceSol?: number | null;
}