# lib/api-zod/src/generated/types/walletInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Svarog - Solana Token Launcher Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface WalletInput {
  ownerAddress: string;
  label?: string;
  group?: string;
}